backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
pillow==9.1.0
pdf2image==1.16.0
werkzeug==2.1.1